requirements.txt
discord.py==2.3.2
openai==0.27.2
zep-python
python-dotenv
tenacity
aiohttp
requests
numpy
asyncpg
watchdog
python-telegram-bot

